#971 [REFACTOR] 안내 모달 저장 위치 변경 및 constant로 수정
