+++ src/feature/event/component/GuideModal/GuideModal.module.css
@@ -25,6 +25,7 @@
   background: var(--blue-0);
   border-radius: 1.5rem;
   width: 90%;
+  min-width: 27.6rem;
   max-width: 30rem;
   height: 36rem;
   padding: 2rem;
